add input value validation
check if the input contains a list at data>attributes>list and
perform some basic checks for the list data, namely date format and
that value/power is an integer. The validation assumes that either
"value" or "power" are acceptable. If neccessary, the input is
casted from string to dict.

Validation errors print an ValueError message and return with 1