sum up values and return sum
this assumes that multiple queries should still be summed up,
resulting in one value field in the return value. The result is
not printed and will only be a return value.

The main function only accepts one command line argument when
called from the command line, I'll write a readme